feat: Add client authentication and commands support
At this moment, there is no public API in Server module to control client authentication. It always allow clients to connect but there is not difficulties to add authentication(see crates\server\src\lib.rs:93).
Client authentication is a need to be able create resource manager. Omgpp-cli should be able access to Resource manager. In order to do that we need some way to communicate with the server without affecting game logic. It could have been accomplished by creating separate TCP socket or simple HTTP server. But decision was made in favor of utilizing already existed GNS socket.
With an authentication mechanism after a connection request we can easily differentiate who is connecting to Server whether it a omgpp-cli or regular player. If it omgpp-cli it can communicate with the server via commands (not messages or RPC)